re-organize classes from SAXSGUIPlugin
